+++ src/gui/widget.rs
@@ -23,9 +23,30 @@ pub use w::Space;
 pub mod id {
     use once_cell::sync::Lazy;
 
-    pub static CUSTOM_GAMES: Lazy<iced::widget::scrollable::Id> = Lazy::new(iced::widget::scrollable::Id::unique);
+    pub static BACKUP_SCROLL: Lazy<iced::widget::scrollable::Id> = Lazy::new(iced::widget::scrollable::Id::unique);
+    pub static RESTORE_SCROLL: Lazy<iced::widget::scrollable::Id> = Lazy::new(iced::widget::scrollable::Id::unique);
+    pub static CUSTOM_GAMES_SCROLL: Lazy<iced::widget::scrollable::Id> =
+        Lazy::new(iced::widget::scrollable::Id::unique);
+    pub static OTHER_SCROLL: Lazy<iced::widget::scrollable::Id> = Lazy::new(iced::widget::scrollable::Id::unique);
+    pub static MODAL_SCROLL: Lazy<iced::widget::scrollable::Id> = Lazy::new(iced::widget::scrollable::Id::unique);
 
-    pub fn custom_games() -> iced::widget::scrollable::Id {
-        (*CUSTOM_GAMES).clone()
+    pub fn backup_scroll() -> iced::widget::scrollable::Id {
+        (*BACKUP_SCROLL).clone()
+    }
+
+    pub fn restore_scroll() -> iced::widget::scrollable::Id {
+        (*RESTORE_SCROLL).clone()
+    }
+
+    pub fn custom_games_scroll() -> iced::widget::scrollable::Id {
+        (*CUSTOM_GAMES_SCROLL).clone()
+    }
+
+    pub fn other_scroll() -> iced::widget::scrollable::Id {
+        (*OTHER_SCROLL).clone()
+    }
+
+    pub fn modal_scroll() -> iced::widget::scrollable::Id {
+        (*MODAL_SCROLL).clone()
     }
 }